requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
boto3>=1.26.0
pytest>=7.0.0
black>=22.0.0
pylint>=2.15.0 